Version 2.0.0.  Works but creates a task per binary component.
Downloaded NAR are extracted in a set of directories named by the
platform-buildType-flavor tuple. The component name is not used
by this version. We still create configurations and tasks for each
binary but should be simplifying this more.

Unit tests are disabled for now, to be rewritten soon with the
new features in place.