Update the plugin to be Gradle 2.9 compatible.
The NativeComponent has to be created eagerly to support resolving
in the publishing plugin as needed. Lazy creation does not work
since the plugin is using an unmanaged project collection to store
the SoftwareComponent instances. This may need to be examined in the
future to better fit the new model but for now I am happy with the result.